+++ tests/broken_plugin.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+"""
+Broken plug-in
+==============
+"""
+
+from iris import AbstractRawDataset
+import _______ as this_is_an_error
+
+class BrokenPlugin(AbstractRawDataset):
+
+    def __init__(self, source=None, metadata=dict()):
+        pass